Moved create new library dialog into a seperate component
